refactor: 경주 게임 준비/실행/종료 에 해당하는 메서드를 모아 클래스로 분리
- CarRacing.java: 경주 게임의 전 과정을 호출하는 클래스
- ReadyToRace.java: 입력 처리 및 객체 생성을 하는 메서드를 담은 클래스
- StartRace.java: 게임을 진행하는 클래스
- FinishRace.java: 우승자를 추출하여 출력하는 메서드를 담은 클래스
- Utils.java: 어느 객체에 속하지 않아도 사용될 수 있는 메서드를 담은 클래스
